+++ tde/widget/settings/calendar.lua
@@ -23,13 +23,10 @@
 --SOFTWARE.
 ]]
 local os = require("os")
-local awful = require("awful")
 local wibox = require("wibox")
-local gears = require("gears")
 local beautiful = require("beautiful")
 local rounded = require("lib-tde.widget.rounded")
 local dpi = beautiful.xresources.apply_dpi
-local icons = require("theme.icons")
 
 local m = dpi(10)
 local settings_index = dpi(40)
@@ -44,22 +41,6 @@ return function()
   local title = wibox.widget.textbox(i18n.translate("Calendar"))
   title.font = beautiful.title_font
   title.forced_height = settings_index + m + m
-
-  local close = wibox.widget.imagebox(icons.close)
-  close.forced_height = settings_index
-  close:buttons(
-    gears.table.join(
-      awful.button(
-        {},
-        1,
-        function()
-          if root.elements.settings then
-            root.elements.settings.close()
-          end
-        end
-      )
-    )
-  )
 
   local cal_container = wibox.container.background()
   cal_container.bg = beautiful.bg_modal
@@ -116,18 +97,6 @@ return function()
     layout = wibox.container.background,
     {
       layout = wibox.layout.align.vertical,
-      {
-        layout = wibox.layout.align.horizontal,
-        nil,
-        wibox.container.margin(
-          {
-            layout = wibox.container.place,
-            title
-          },
-          settings_index * 2
-        ),
-        close
-      },
       {
         layout = wibox.container.place,
         valign = "top",